Reorganize project structure with src folder. Endpoints and pytest still work. Still need dev.db in src folder at this point.
